+++ spec/examples/usual/dynamic_options/format/date/message/static/example2_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Usual::DynamicOptions::Format::Date::Message::Example1 do
+RSpec.describe Usual::DynamicOptions::Format::Date::Message::Static::Example2 do
   describe ".call!" do
     subject(:perform) { described_class.call!(**attributes) }
 
@@ -14,7 +14,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Message::Example1 do
 
     include_examples "check class info",
                      inputs: %i[started_on],
-                     internals: %i[],
+                     internals: %i[started_on],
                      outputs: %i[started_on]
 
     context "when the input arguments are valid" do
@@ -34,7 +34,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Message::Example1 do
           it "returns expected error" do
             expect { perform }.to(
               raise_error(
-                ApplicationService::Exceptions::Input,
+                ApplicationService::Exceptions::Internal,
                 "Invalid date format"
               )
             )
@@ -65,7 +65,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Message::Example1 do
 
     include_examples "check class info",
                      inputs: %i[started_on],
-                     internals: %i[],
+                     internals: %i[started_on],
                      outputs: %i[started_on]
 
     context "when the input arguments are valid" do
@@ -85,7 +85,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Message::Example1 do
           it "returns expected error" do
             expect { perform }.to(
               raise_error(
-                ApplicationService::Exceptions::Input,
+                ApplicationService::Exceptions::Internal,
                 "Invalid date format"
               )
             )
